refactor(src/game): game files polishing
