--- packages/client/src/game/classesGame/Game/index.ts
@@ -1,3 +1,0 @@
-import Game from './Game'
-
-export { Game as GameEngine }
