--- packages/client/src/game/classesGame/Input/index.ts
@@ -1,3 +1,0 @@
-import Input from './Input'
-
-export {Input as InputHandler}
