--- packages/client/src/game/classesGame/Sprite/index.ts
@@ -1,3 +1,0 @@
-import Sprite from './Sprite'
-
-export default Sprite
