--- packages/client/src/game/classesGame/Sprite/types.ts
@@ -1,8 +1,0 @@
-import Sprite from './Sprite'
-
-export interface ISprite {
-  image: HTMLImageElement;
-  src: string;
-
-  load(): Promise<Sprite>;
-}
