--- packages/client/src/game/classesGame/View/index.ts
@@ -1,3 +1,0 @@
-import View from './View'
-
-export default View
