--- packages/client/src/game/classesGame/Wall/types.ts
@@ -1,3 +1,0 @@
-import { GameObjectArgs } from '@/game/classesGame/GameObject/types'
-
-export type WallArgs = {type: undefined | string} & GameObjectArgs
