+++ packages/client/src/game/core/game-engine/index.ts
@@ -1,0 +1,3 @@
+import GameEngine from './game-engine'
+
+export default GameEngine
