+++ packages/client/src/game/core/input/index.ts
@@ -1,0 +1,3 @@
+import Input from './input'
+
+export { Input as InputHandler }
