+++ packages/client/src/game/core/sprite/index.ts
@@ -1,0 +1,3 @@
+import Sprite from './sprite'
+
+export default Sprite
