+++ packages/client/src/game/core/sprite/types.ts
@@ -1,0 +1,8 @@
+import Sprite from './sprite'
+
+export interface ISprite {
+  image: HTMLImageElement;
+  src: string;
+
+  load(): Promise<Sprite>;
+}
