+++ packages/client/src/game/core/view/index.ts
@@ -1,0 +1,3 @@
+import View from './view'
+
+export default View
