+++ packages/client/src/game/core/wall/types.ts
@@ -1,0 +1,3 @@
+import { GameObjectArgs } from '@/game/core/types'
+
+export type WallArgs = {type: undefined | string} & GameObjectArgs
