+++ packages/client/src/game/helpers/types.ts
@@ -1,0 +1,3 @@
+export type Sprites = [number, number, number, number][]
+
+export type DefaultPosition = [number, number]
