+++ packages/client/src/game/sprites/originalSprite.png
