+++ packages/client/src/game/sprites/sprite_1.png
